Move get_nav_data into cbv.queries
